Improvements to CI to make it less flaky.

+++ tests/tester.nim
@@ -1,5 +1,8 @@
 import asynctools, asyncdispatch, os, httpclient, strutils, asyncnet
 
+from unittest import check, abortOnError
+abortOnError = true
+
 from osproc import execCmd
 
 var serverProcess: AsyncProcess
@@ -29,6 +32,13 @@ proc startServer(file: string) {.async.} =
     doAssert execCmd("kill -15 " & $serverProcess.processID()) == QuitSuccess
     serverProcess = nil
 
+  # Verify that another server isn't already running.
+  doAssertRaises(OSError):
+    let client = newAsyncHttpClient()
+    let data = await client.getContent("http://localhost:8080/")
+    echo("Got data from server: ", data)
+    echo("Received data from server, which means that stale serve is running.")
+
   # The nim process doesn't behave well when using `-r`, if we kill it, the
   # process continues running...
   doAssert execCmd("nimble c -y " & file) == QuitSuccess
@@ -44,9 +54,9 @@ proc tests() {.async.} =
   block:
     let client = newAsyncHttpClient()
     let resp = await client.get("http://localhost:8080/")
-    doAssert resp.code == Http200
+    check resp.code == Http200
     let body = await resp.body
-    doAssert body == "Hello World"
+    check body == "Hello World"
 
   await startServer("dispatcher.nim")
 
@@ -54,17 +64,17 @@ proc tests() {.async.} =
   block:
     let client = newAsyncHttpClient()
     let resp = await client.get("http://localhost:8080")
-    doAssert resp.code == Http200
+    check resp.code == Http200
     let body = await resp.body
-    doAssert body == "Hi there!"
+    check body == "Hi there!"
 
   # Simple POST
   block:
     let client = newAsyncHttpClient()
     let resp = await client.post("http://localhost:8080", body="hello")
-    doAssert resp.code == Http200
+    check resp.code == Http200
     let body = await resp.body
-    doAssert body == "Successful POST! Data=5"
+    check body == "Successful POST! Data=5"
 
   await startServer("simpleLog.nim")
 
@@ -73,16 +83,16 @@ proc tests() {.async.} =
   block:
     let client = newAsyncHttpClient()
     let resp = await client.get("http://localhost:8080")
-    doAssert resp.code == Http200
-    doAssert logFilename.readLines() == @["INFO Requested /"]
+    check resp.code == Http200
+    check logFilename.readLines() == @["INFO Requested /"]
 
   block:
     let client = newAsyncHttpClient()
     let resp = await client.get("http://localhost:8080/404")
-    doAssert resp.code == Http404
-    doAssert logFilename.readLines(2) == @["INFO Requested /", "ERROR 404"]
+    check resp.code == Http404
+    check logFilename.readLines(2) == @["INFO Requested /", "ERROR 404"]
 
-  doAssert tryRemoveFile(logFilename)
+  check tryRemoveFile(logFilename)
 
   # Verify cross-talk doesn't occur
   await startServer("crosstalk.nim")
@@ -97,11 +107,11 @@ proc tests() {.async.} =
     await client.connect("localhost", Port(8080))
     await client.send("GET /2 HTTP/1.1\c\l\c\l")
 
-    doAssert (await client.recvLine()) == "HTTP/1.1 200 OK"
-    doAssert (await client.recvLine()) == "Content-Length: 10"
-    doAssert (await client.recvLine()).startsWith("Server")
-    doAssert (await client.recvLine()).startsWith("Date:")
-    doAssert (await client.recvLine()) == "\c\l"
+    check (await client.recvLine()) == "HTTP/1.1 200 OK"
+    check (await client.recvLine()) == "Content-Length: 10"
+    check (await client.recvLine()).startsWith("Server")
+    check (await client.recvLine()).startsWith("Date:")
+    check (await client.recvLine()) == "\c\l"
     let delayedBody = await client.recv(10)
     doAssert(delayedBody == "Delayed /2", "We must get the ID we asked for.")
 
@@ -109,18 +119,18 @@ proc tests() {.async.} =
     var client = newAsyncSocket()
     await client.connect("localhost", Port(8080))
     await client.send("GET /close_me/1 HTTP/1.1\c\l\c\l")
-    doAssert (await client.recv(1)) == ""
+    check (await client.recv(1)) == ""
     client.close()
 
     client = newAsyncSocket()
     defer: client.close()
     await client.connect("localhost", Port(8080))
     await client.send("GET /close_me/2 HTTP/1.1\c\l\c\l")
-    doAssert (await client.recvLine()) == "HTTP/1.1 200 OK"
-    doAssert (await client.recvLine()) == "Content-Length: 19"
-    doAssert (await client.recvLine()).startsWith("Server")
-    doAssert (await client.recvLine()).startsWith("Date:")
-    doAssert (await client.recvLine()) == "\c\l"
+    check (await client.recvLine()) == "HTTP/1.1 200 OK"
+    check (await client.recvLine()) == "Content-Length: 19"
+    check (await client.recvLine()).startsWith("Server")
+    check (await client.recvLine()).startsWith("Date:")
+    check (await client.recvLine()) == "\c\l"
     let delayedBody = await client.recv(19)
     doAssert(delayedBody == "Delayed /close_me/2", "We must get the ID we asked for.")
 
@@ -130,4 +140,5 @@ when isMainModule:
   try:
     waitFor tests()
   finally:
-    doAssert execCmd("kill -15 " & $serverProcess.processID()) == QuitSuccess
+    if not serverProcess.isNil:
+      doAssert execCmd("kill -15 " & $serverProcess.processID()) == QuitSuccess
